Move some header definitions to compare_internal.cc
