+++ cpp/src/arrow/meson.build
@@ -132,6 +132,7 @@ arrow_components = {
             'chunked_array.cc',
             'chunk_resolver.cc',
             'compare.cc',
+            'compare_internal.cc',
             'config.cc',
             'datum.cc',
             'device.cc',
